+++ src/modules/Chat/GuideModal/index.tsx
@@ -6,26 +6,26 @@ import dynamic from "next/dynamic";
 
 import { useToggle } from "src/hooks";
 import { Box, Button, CheckBox } from "src/ui";
-import { AlertDialogContent } from "src/ui/Popup/Popup";
+import { AlertDialogContent } from "src/ui/Popup/Contents";
 
 import StepIndicator from "./StepIndicator";
 import Step1 from "./steps/Step1";
 
 export const guideStorage = reastorage("neverShowGuideAgain", false);
 
-const STEP_MAP = {
-  0: Step1,
-  1: dynamic(() => import("./steps/Step2"), { suspense: true }),
-  2: dynamic(() => import("./steps/Step3"), { suspense: true }),
-  3: dynamic(() => import("./steps/Step4"), { suspense: true }),
-  4: dynamic(() => import("./steps/Step5"), { suspense: true }),
-} as const;
+const STEP_MAP = [
+  Step1,
+  dynamic(() => import("./steps/Step2"), { suspense: true }),
+  dynamic(() => import("./steps/Step3"), { suspense: true }),
+  dynamic(() => import("./steps/Step4"), { suspense: true }),
+  dynamic(() => import("./steps/Step5"), { suspense: true }),
+] as const;
 
-const TOTAL_STEPS = Object.keys(STEP_MAP).length;
+const TOTAL_STEPS = STEP_MAP.length;
 
 const GuideModal = () => {
   const setNeverShowAgain = useSetReastorage(guideStorage);
-  const [step, setStep] = useState<keyof typeof STEP_MAP>(0);
+  const [step, setStep] = useState(0);
   const StepComponent = STEP_MAP[step];
 
   const isLastStep = TOTAL_STEPS - 1 <= step;
@@ -34,9 +34,7 @@ const GuideModal = () => {
     startTransition(() => {
       setStep((prevStep) => {
         const nextStep = prevStep + stepChangeAmount;
-        return (
-          nextStep < 0 || nextStep >= TOTAL_STEPS ? prevStep : nextStep
-        ) as keyof typeof STEP_MAP;
+        return nextStep < 0 || nextStep >= TOTAL_STEPS ? prevStep : nextStep;
       });
     });
   };
